Backend(python)/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sib-api-v3-sdk